top: rework RCW source selection
We don't need the power off special case for
CPLD_INTERRUPT_CFG_RCW_SRC2. Normally, the pin is already open drain,
thus in this case we don't need to worry about that pin driving an
unpowered input. Only if FORCE_RECOV# is active, we might drive an high
value. But this is impossible when the board is off, because then the
PORESET# is active, which leads to irq_out being 0.

Signed-off-by: Michael Walle <michael@walle.cc>