top: fix power off request handling
Do not evaluate the CPLD_INTERRUPT_CFG_RCW_SRC2 pin in failsafe mode
because that might cause undesired power-off. This is because we might
be in test mode where there is no pull-up and the low level of the
signal will be interpreted as a power-off request on each reset.

Signed-off-by: Michael Walle <michael@walle.cc>